+++ src/itables/widget/__init__.py
@@ -26,12 +26,7 @@ class ITable(anywidget.AnyWidget):
 
     # private traits that relate to df or to the DataTable arguments
     # (use .update() to update them)
-    _data_json = traitlets.Unicode().tag(sync=True)
-    _columns = traitlets.List(traitlets.Dict()).tag(sync=True)
-    _filtered_row_count = traitlets.Int().tag(sync=True)
-    _downsampling_warning = traitlets.Unicode().tag(sync=True)
     _dt_args = traitlets.Dict().tag(sync=True)
-    _destroy_and_recreate = traitlets.Int(0).tag(sync=True)
 
     def __init__(
         self,
@@ -47,11 +42,7 @@ class ITable(anywidget.AnyWidget):
         self.style = other_args.pop("style")
         self.selected_rows = other_args.pop("selected_rows")
 
-        self._data_json = dt_args.pop("data_json")
-        self._columns = dt_args.pop("columns")
         self._dt_args = dt_args
-        self._downsampling_warning = other_args.pop("downsampling_warning") or ""
-        self._filtered_row_count = other_args.pop("filtered_row_count", 0)
         assert not other_args, other_args
 
     def update(
@@ -64,12 +55,13 @@ class ITable(anywidget.AnyWidget):
         Update either the table data, attributes, or the arguments passed
         to DataTable. Arguments that are not mentioned
         """
-        data_or_dt_args_changed = False
+        dt_args_changed = False
+        new_dt_args = self._dt_args.copy()
         for key, value in list(kwargs.items()):
             if value is None:
-                data_or_dt_args_changed = True
-                self._dt_args.pop(key, None)
+                dt_args_changed = True
                 del kwargs[key]
+                del new_dt_args[key]
 
         if df is None:
             df = self._df
@@ -91,33 +83,21 @@ class ITable(anywidget.AnyWidget):
         if df is None:
             del dt_args["data_json"]
             del dt_args["columns"]
-
-            # Don't trigger an update if nor data nor the dt args changed
-            data_or_dt_args_changed = data_or_dt_args_changed or self._update_dt_args(
-                dt_args
-            )
+            del dt_args["filtered_row_count"]
+            del dt_args["downsampling_warning"]
         else:
             self._df = df
-            self._data_json = dt_args.pop("data_json")
-            self._columns = dt_args.pop("columns")
-            self._update_dt_args(dt_args)
-            self._downsampling_warning = other_args.pop("downsampling_warning") or ""
-            self._filtered_row_count = other_args.pop("filtered_row_count", 0)
-            data_or_dt_args_changed = True
+            dt_args_changed = True
 
-        if data_or_dt_args_changed:
-            self._destroy_and_recreate += 1
+        for key, value in dt_args.items():
+            if key not in new_dt_args or (new_dt_args[key] != value):
+                new_dt_args[key] = value
+                dt_args_changed = True
+
+        if dt_args_changed:
+            self._dt_args = new_dt_args
 
         self.selected_rows = other_args.pop("selected_rows")
-
-    def _update_dt_args(self, dt_args):
-        changed = False
-        for key, value in dt_args.items():
-            if key not in self._dt_args or (self._dt_args[key] != value):
-                self._dt_args[key] = value
-                changed = True
-
-        return changed
 
     @property
     def df(self):