+++ src/itables_for_dash/properties.py
@@ -18,15 +18,11 @@ from itables.javascript import get_itables_extension_arguments
 from itables.typing import DataTableOptions, ITableOptions
 
 ITABLE_PROPERTIES = (
-    "data_json",
-    "columns",
     "caption",
-    "selected_rows",
     "style",
     "classes",
     "dt_args",
-    "filtered_row_count",
-    "downsampling_warning",
+    "selected_rows",
 )
 
 
@@ -43,8 +39,6 @@ def get_itable_component_kwargs(
     style["tableLayout"] = style.pop("table-layout")
 
     return {
-        "data_json": dt_args.pop("data_json"),
-        "columns": dt_args.pop("columns"),
         "dt_args": dt_args,
         "style": style,
         **other_args,
@@ -62,14 +56,17 @@ def updated_itable_outputs(
 ):
     updated_properties = get_itable_component_kwargs(df, **kwargs)
 
-    if df is None:
-        updated_properties["data_json"] = no_update
-        updated_properties["columns"] = no_update
-        updated_properties["filtered_row_count"] = no_update
-        updated_properties["downsampling_warning"] = no_update
-        updated_properties["selected_rows"] = kwargs.get("selected_rows") or []
-
     if current_dt_args is not None:
+        if df is None:
+            for k in {
+                "columns",
+                "data_json",
+                "filtered_row_count",
+                "downsampling_warning",
+            }:
+                if k in current_dt_args:
+                    updated_properties["dt_args"][k] = current_dt_args[k]
+
         if current_dt_args == updated_properties["dt_args"]:
             updated_properties["dt_args"] = no_update